One data structure per file
