+++ TrackerType.swift
@@ -1,0 +1,16 @@
+//
+//  JustTrack
+//
+//  Copyright © 27/12/2021 Just Eat Holding Ltd.
+//
+
+import Foundation
+
+/// Tracker type. Used for trackers provided by JustTrack out of the box.
+///
+/// ````
+/// case consoleLogger
+/// ````
+public enum TrackerType {
+    case consoleLogger
+}
